Split the saved links class into its own file, for easier testing.
